Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/bufferedAmount/bufferedAmount-arraybuffer-expected.txt


FAIL WebSockets: bufferedAmount for ArrayBuffer assert_equals: expected 10 but got 0
